+++ theme/template/front-page.php
@@ -1,5 +1,19 @@
 <?php get_header(); ?>
 
+<?php
+// Generating add-to-calendar link.
+$calendar = implode([
+  'https://www.google.com/calendar/event?action=TEMPLATE',
+  'text=JP_Stripes+Connect+2020',
+  'dates=20200418T010000Z/20200418T100000Z',
+  'location='.urlencode('東京都港区南青山3-8-38 南青山東急ビル'),
+  'details='.urlencode(implode([
+    'Website',
+    'https://connect2020.jpstripes.com',
+  ], "\n")),
+], '&');
+?>
+
 <header class="header header--home">
   <div class="container">
     <div class="row">
@@ -21,9 +35,19 @@
             10:00-19:00
           </p>
 
-          <a class="btn btn-lg btn-danger display--inline" href="#">
-            チケット申し込み
-          </a>
+          <div class="mb-4">
+            <a class="btn btn-lg btn-danger display--inline" target="_blank" href="#">
+              &nbsp;&nbsp;
+              チケット申し込み
+              &nbsp;&nbsp;
+            </a>
+          </div>
+
+          <div>
+            <a class="home-tagline__calendar" target="_blank" href="<?= $calendar ?>">
+              Googleカレンダーに登録
+            </a>
+          </div>
         </section>
       </div>
     </div>